Minor optimizations, add release build variant

+++ Src/Nes/Bus.h
@@ -26,7 +26,7 @@ typedef struct _DMA_t
 {
   DMA_State_t State;
   uint16_t CPUBaseAddress;
-  uint16_t NumTransfersComplete;
+  uint8_t NumTransfersComplete;
   uint8_t Data;
 } DMA_t;
